Add files via upload
This program is a CircuitPython program to play the Tic-Tac-Toe on an Adafruit Macropad computer. Instructions for the game are displayed on the built-in screen, and are fairly straightforward.

This program does require a number of Adafruit libraries. they are not included here as I have not checked licensing restrictions.
More details are available ahttps://learn.adafruit.com/adafruit-macropad-rp2040/macropad-circuitpython-library